--- scuttlebutt-rpc/src/integrationTest/java/org/apache/tuweni/scuttlebutt/rpc/PatchworkIntegrationTest.java
@@ -1,166 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE
- * file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file
- * to You under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
- * License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package org.apache.tuweni.scuttlebutt.rpc;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.apache.tuweni.bytes.Bytes;
-import org.apache.tuweni.bytes.Bytes32;
-import org.apache.tuweni.concurrent.AsyncResult;
-import org.apache.tuweni.crypto.sodium.Signature;
-import org.apache.tuweni.io.Base64;
-import org.apache.tuweni.junit.VertxExtension;
-import org.apache.tuweni.junit.VertxInstance;
-import org.apache.tuweni.scuttlebutt.handshake.vertx.ClientHandler;
-import org.apache.tuweni.scuttlebutt.handshake.vertx.SecureScuttlebuttVertxClient;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Scanner;
-import java.util.function.Consumer;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.vertx.core.Vertx;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-/**
- * Test used with a local installation of Patchwork on the developer machine.
- * <p>
- * Run patchwork with: docker run --rm -p 7777:7777 -p 8008:8008 -v ~/.ssb:/home/node/.ssb --name patchwork
- * mazzolino/patchwork
- */
-@ExtendWith(VertxExtension.class)
-class PatchworkIntegrationTest {
-
-  public static class MyClientHandler implements ClientHandler {
-
-    private final Consumer<Bytes> sender;
-    private final Runnable terminationFn;
-
-    public MyClientHandler(Consumer<Bytes> sender, Runnable terminationFn) {
-      this.sender = sender;
-      this.terminationFn = terminationFn;
-    }
-
-    @Override
-    public void receivedMessage(Bytes message) {
-
-      RPCMessage rpcMessage = new RPCMessage(message);
-
-      System.out.println(rpcMessage.asString());
-    }
-
-    @Override
-    public void streamClosed() {
-      System.out.println("Stream closed?");
-
-    }
-
-    void sendMessage(Bytes bytes) {
-      System.out.println("Sending message?");
-      sender.accept(bytes);
-    }
-  }
-
-  /**
-   * This test tests the connection to a local patchwork installation. You need to run patchwork locally to perform that
-   * work.
-   *
-   */
-  @Test
-  void runWithPatchWork(@VertxInstance Vertx vertx) throws Exception {
-    String host = "localhost";
-    int port = 8008;
-
-    Optional<String> ssbDir = Optional.ofNullable(System.getenv().get("ssb_dir"));
-    Optional<String> homePath = Optional.ofNullable(System.getProperty("user.home")).map(home -> home + "/.ssb");
-
-    Optional<String> path = Optional.of(ssbDir.orElseGet(homePath::get));
-
-    String secretPath = path.get() + "/secret";
-    File file = new File(secretPath);
-
-    if (!file.exists()) {
-      throw new Exception("Secret file does not exist");
-    }
-
-    Scanner s = new Scanner(file, UTF_8.name());
-    s.useDelimiter("\n");
-
-    ArrayList<String> list = new ArrayList<>();
-    while (s.hasNext()) {
-      String next = s.next();
-
-      // Filter out the comment lines
-      if (!next.startsWith("#")) {
-        list.add(next);
-      }
-    }
-
-    String secretJSON = String.join("", list);
-
-    ObjectMapper mapper = new ObjectMapper();
-
-    HashMap<String, String> values = mapper.readValue(secretJSON, new TypeReference<Map<String, String>>() {});
-    String pubKey = values.get("public").replace(".ed25519", "");
-    String privateKey = values.get("private").replace(".ed25519", "");
-
-    Bytes pubKeyBytes = Base64.decode(pubKey);
-    Bytes privKeyBytes = Base64.decode(privateKey);
-
-    Signature.PublicKey pub = Signature.PublicKey.fromBytes(pubKeyBytes);
-    Signature.SecretKey secretKey = Signature.SecretKey.fromBytes(privKeyBytes);
-
-    Signature.KeyPair keyPair = new Signature.KeyPair(pub, secretKey);
-    String networkKeyBase64 = "1KHLiKZvAvjbY1ziZEHMXawbCEIM6qwjCDm3VYRan/s=";
-
-    String serverPublicKey = pubKey;
-    Signature.PublicKey publicKey = Signature.PublicKey.fromBytes(Base64.decode(serverPublicKey));
-
-    Bytes32 networkKeyBytes32 = Bytes32.wrap(Base64.decode(networkKeyBase64));
-
-    SecureScuttlebuttVertxClient secureScuttlebuttVertxClient =
-        new SecureScuttlebuttVertxClient(vertx, keyPair, networkKeyBytes32);
-
-    AsyncResult<ClientHandler> onConnect =
-        secureScuttlebuttVertxClient.connectTo(port, host, publicKey, MyClientHandler::new);
-
-    MyClientHandler clientHandler = (MyClientHandler) onConnect.get();
-    assertTrue(onConnect.isDone());
-    assertFalse(onConnect.isCompletedExceptionally());
-    Thread.sleep(1000);
-    assertNotNull(clientHandler);
-    // An RPC command that just tells us our public key (like ssb-server whoami on the command line.)
-    String rpcRequestBody = "{\"name\": [\"whoami\"],\"type\": \"async\",\"args\":[]}";
-    Bytes rpcRequest = RPCCodec.encodeRequest(rpcRequestBody, RPCFlag.BodyType.JSON);
-
-    System.out.println("Attempting RPC request...");
-    clientHandler.sendMessage(rpcRequest);
-    for (int i = 0; i < 10; i++) {
-      clientHandler.sendMessage(RPCCodec.encodeRequest(rpcRequestBody, RPCFlag.BodyType.JSON));
-    }
-
-    Thread.sleep(10000);
-
-    secureScuttlebuttVertxClient.stop().join();
-  }
-
-}
